Why do you have to make this so painful
Uda "let's do simple things in a complicated way" city. We could just use a shared ptr in ChatBot _image and be done with it, but instead I am supposed to write 50 lines of useless manual memory management code to do the same thing, but slower, and more error-prone. Sigh...